Add mutex lock for App,Node,Queue daoSnapshot() funcs.
In the unit-tests (and regular usage), there may be concurrent
goroutines that are emitting events to the event-stream, and those
include the DAO state generated by daoSnapshot() on these objects - to
prevent deadlock/races, add a fine-grained mutex just for the DAO
snapshot generation. Also, remove a flaky assertion for checking
partition name on a node - it may not yet be assigned.